requirements.txt
python-telegram-bot==12.7
psycopg2-binary==2.9.3
psycopg2==2.9.3